+++ source/server/Karamem0.Commistant.Common/ConfigureServices.cs
@@ -6,12 +6,20 @@
 // https://github.com/karamem0/commistant/blob/main/LICENSE
 //
 
+using Azure.Identity;
+using Azure.Storage;
+using Karamem0.Commistant.Adapters;
 using Karamem0.Commistant.Options;
 using Karamem0.Commistant.Services;
-using Microsoft.Agents.Authentication;
-using Microsoft.Agents.Builder;
+using Microsoft.Agents.Builder.Adapters;
+using Microsoft.Agents.Builder.App;
+using Microsoft.Agents.Builder.State;
+using Microsoft.Agents.Hosting.AspNetCore;
+using Microsoft.Agents.Storage;
+using Microsoft.Agents.Storage.Blobs;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using OpenAI;
 using QRCoder;
 
@@ -24,15 +32,35 @@ public static class ConfigureServices
     {
         _ = services.Configure<AzureStorageBlobsOptions>(configuration.GetSection("AzureStorageBlobs"));
         _ = services.Configure<AzureOpenAIOptions>(configuration.GetSection("AzureOpenAI"));
-        _ = services.Configure<ConnectorClientOptions>(configuration.GetSection("ConnectorClient"));
         return services;
     }
 
-    public static IServiceCollection AddConnectorClient(this IServiceCollection services)
+    public static void AddAgent<T>(this IHostApplicationBuilder builder, IConfiguration configuration) where T : AgentApplication
     {
-        _ = services.AddSingleton<IConnections, ConfigurationConnections>();
-        _ = services.AddSingleton<IChannelServiceClientFactory, RestChannelServiceClientFactory>();
-        return services;
+        _ = builder.AddAgent<T, AdapterWithErrorHandler>();
+        _ = builder.Services.AddSingleton((provider) => new AgentApplicationOptions(provider.GetRequiredService<IStorage>())
+            {
+                ChannelAdapterRegistry = provider.GetRequiredService<IChannelAdapterRegistry>(),
+                TurnStateFactory = () => new TurnState(
+                    provider.GetRequiredService<ConversationState>(),
+                    provider.GetRequiredService<UserState>(),
+                    new TempState()
+                )
+            }
+        );
+        var options = configuration
+            .GetSection("AzureStorageBlobs")
+            .Get<AzureStorageBlobsOptions>();
+        _ = options ?? throw new InvalidOperationException($"{nameof(AzureStorageBlobsOptions)} を null にはできません");
+        _ = builder.Services.AddSingleton<IStorage>(
+            new BlobsStorage(
+                new Uri(options.Endpoint, options.ContainerName),
+                new DefaultAzureCredential(new DefaultAzureCredentialOptions()),
+                new StorageTransferOptions()
+            )
+        );
+        _ = builder.Services.AddSingleton<ConversationState>();
+        _ = builder.Services.AddSingleton<UserState>();
     }
 
     public static IServiceCollection AddServices(this IServiceCollection services, IConfiguration configuration)