+++ source/server/Karamem0.Commistant.Common/Services/ConnectorClientService.cs
@@ -6,14 +6,13 @@
 // https://github.com/karamem0/commistant/blob/main/LICENSE
 //
 
-using Karamem0.Commistant.Options;
+using Microsoft.Agents.Authentication;
 using Microsoft.Agents.Builder;
-using Microsoft.Agents.Connector;
+using Microsoft.Agents.Builder.App;
+using Microsoft.Agents.Builder.App.Proactive;
 using Microsoft.Agents.Core.Models;
 using Microsoft.Agents.Extensions.Teams.Connector;
 using Microsoft.Agents.Extensions.Teams.Models;
-using Microsoft.Extensions.Options;
-using System.Security.Claims;
 using System.Threading;
 
 namespace Karamem0.Commistant.Services;
@@ -21,71 +20,84 @@ namespace Karamem0.Commistant.Services;
 public interface IConnectorClientService
 {
 
-    Task<ResourceResponse> SendActivityAsync(
-        string serviceUrl,
+    Task SendActivityAsync(
+        ConversationReference conversationReference,
         IActivity activity,
         CancellationToken cancellationToken = default
     );
 
     Task<MeetingInfo> GetMeetingInfoAsync(
-        string serviceUrl,
+        ConversationReference conversationReference,
         string meetingId,
         CancellationToken cancellationToken = default
     );
 
 }
 
-public class ConnectorClientService(IChannelServiceClientFactory factory, IOptions<ConnectorClientOptions> options) : IConnectorClientService
+public class ConnectorClientService(IAgent agent, IConnections connections) : IConnectorClientService
 {
 
-    private readonly IChannelServiceClientFactory factory = factory;
+    private readonly AgentApplication agent = (AgentApplication)agent;
 
-    private readonly ConnectorClientOptions options = options.Value;
+    private readonly IConnections connections = connections;
 
-    [Obsolete]
-    public async Task<ResourceResponse> SendActivityAsync(
-        string serviceUrl,
+    public async Task SendActivityAsync(
+        ConversationReference conversationReference,
         IActivity activity,
         CancellationToken cancellationToken = default
     )
     {
-        var claimsIdentity = new ClaimsIdentity(
-            [
-                new Claim("ver", "1.0"),
-                new Claim("appid", this.options.ClientId)
-            ]
+        var connection = this.connections.GetDefaultConnection();
+        var conversation = new Conversation(
+            AgentClaims.CreateIdentity(connection.ConnectionSettings.ClientId),
+            conversationReference
         );
-        var connectorClient = await this.factory.CreateConnectorClientAsync(
-            claimsIdentity,
-            serviceUrl,
-            this.options.Audience,
-            cancellationToken: cancellationToken
+        _ = await this.agent.Proactive.SendActivityAsync(
+            conversation,
+            activity,
+            cancellationToken
         );
-        return await connectorClient.Conversations.SendToConversationAsync(activity, cancellationToken);
     }
 
-    [Obsolete]
     public async Task<MeetingInfo> GetMeetingInfoAsync(
-        string serviceUrl,
+        ConversationReference conversationReference,
         string meetingId,
         CancellationToken cancellationToken = default
     )
     {
-        var claimsIdentity = new ClaimsIdentity(
-            [
-                new Claim("ver", "1.0"),
-                new Claim("appid", this.options.ClientId)
-            ]
+        var completionSource = new TaskCompletionSource<MeetingInfo>(TaskCreationOptions.RunContinuationsAsynchronously);
+        var connection = this.connections.GetDefaultConnection();
+        var conversation = new Conversation(
+            AgentClaims.CreateIdentity(connection.ConnectionSettings.ClientId),
+            conversationReference
         );
-        var connectorClient = await this.factory.CreateConnectorClientAsync(
-            claimsIdentity,
-            serviceUrl,
-            this.options.Audience,
+        var base64 = Convert.ToBase64String(Encoding.UTF8.GetBytes($"0#{meetingId}#0"));
+        await this.agent.Proactive.ContinueConversationAsync(
+            conversation,
+            async (
+                turnContext,
+                turnState,
+                cancellationToken
+            ) =>
+            {
+                try
+                {
+                    _ = completionSource.TrySetResult(
+                        await TeamsInfo.GetMeetingInfoAsync(
+                            turnContext,
+                            base64,
+                            cancellationToken
+                        )
+                    );
+                }
+                catch (Exception ex)
+                {
+                    _ = completionSource.TrySetException(ex);
+                }
+            },
             cancellationToken: cancellationToken
         );
-        var teamsConnectorClient = new RestTeamsConnectorClient(connectorClient, (IRestTransport)connectorClient);
-        var base64 = Convert.ToBase64String(Encoding.UTF8.GetBytes($"0#{meetingId}#0"));
-        return await teamsConnectorClient.Teams.FetchMeetingInfoAsync(base64, cancellationToken: cancellationToken);
+        return await completionSource.Task.WaitAsync(cancellationToken);
     }
 
 }