+++ source/server/Karamem0.Commistant.Function/Agents/FunctionAgentApplication.cs
@@ -1,0 +1,13 @@
+//
+// Copyright (c) 2022-2026 karamem0
+//
+// This software is released under the MIT License.
+//
+// https://github.com/karamem0/commistant/blob/main/LICENSE
+//
+
+using Microsoft.Agents.Builder.App;
+
+namespace Karamem0.Commistant.Agents;
+
+public class FunctionAgentApplication(AgentApplicationOptions options) : AgentApplication(options);