+++ source/server/Karamem0.Commistant.Function/Commands/MeetingInProgressCommand.cs
@@ -83,8 +83,8 @@ public class MeetingInProgressCommand(
             activity.From = conversationReference.Agent;
             activity.Recipient = conversationReference.User;
             activity.Conversation = conversationReference.Conversation;
-            _ = await this.connectorClientService.SendActivityAsync(
-                conversationReference.ServiceUrl,
+            await this.connectorClientService.SendActivityAsync(
+                conversationReference,
                 activity,
                 cancellationToken
             );