+++ source/server/Karamem0.Commistant.Test/Commands/MeetingInProgressCommandTests.cs
@@ -44,9 +44,6 @@ public class MeetingInProgressCommandTests
             MeetingInProgressUrl = "https://www.example.com/",
         };
         var connectorClientService = Substitute.For<IConnectorClientService>();
-        _ = connectorClientService
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>())
-            .Returns(new ResourceResponse());
         var dateTimeService = Substitute.For<IDateTimeService>();
         _ = dateTimeService
             .GetCurrentDateTime()
@@ -70,7 +67,7 @@ public class MeetingInProgressCommandTests
         // Assert
         _ = connectorClientService
             .Received()
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>());
+            .SendActivityAsync(conversationReference, Arg.Any<Activity>());
         _ = qrCodeService
             .Received()
             .CreateAsync("https://www.example.com/");
@@ -98,9 +95,6 @@ public class MeetingInProgressCommandTests
             MeetingInProgressUrl = "https://www.example.com/",
         };
         var connectorClientService = Substitute.For<IConnectorClientService>();
-        _ = connectorClientService
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>())
-            .Returns(new ResourceResponse());
         var dateTimeService = Substitute.For<IDateTimeService>();
         _ = dateTimeService
             .GetCurrentDateTime()
@@ -124,7 +118,7 @@ public class MeetingInProgressCommandTests
         // Assert
         _ = connectorClientService
             .DidNotReceive()
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>());
+            .SendActivityAsync(conversationReference, Arg.Any<Activity>());
         _ = qrCodeService
             .DidNotReceive()
             .CreateAsync("https://www.example.com/");
@@ -152,9 +146,6 @@ public class MeetingInProgressCommandTests
             MeetingInProgressUrl = "https://www.example.com/",
         };
         var connectorClientService = Substitute.For<IConnectorClientService>();
-        _ = connectorClientService
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>())
-            .Returns(new ResourceResponse());
         var dateTimeService = Substitute.For<IDateTimeService>();
         _ = dateTimeService
             .GetCurrentDateTime()
@@ -178,7 +169,7 @@ public class MeetingInProgressCommandTests
         // Assert
         _ = connectorClientService
             .DidNotReceive()
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>());
+            .SendActivityAsync(conversationReference, Arg.Any<Activity>());
         _ = qrCodeService
             .DidNotReceive()
             .CreateAsync("https://www.example.com/");
@@ -206,9 +197,6 @@ public class MeetingInProgressCommandTests
             MeetingInProgressUrl = "https://www.example.com/",
         };
         var connectorClientService = Substitute.For<IConnectorClientService>();
-        _ = connectorClientService
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>())
-            .Returns(new ResourceResponse());
         var dateTimeService = Substitute.For<IDateTimeService>();
         _ = dateTimeService
             .GetCurrentDateTime()
@@ -232,7 +220,7 @@ public class MeetingInProgressCommandTests
         // Assert
         _ = connectorClientService
             .DidNotReceive()
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>());
+            .SendActivityAsync(conversationReference, Arg.Any<Activity>());
         _ = qrCodeService
             .DidNotReceive()
             .CreateAsync("https://www.example.com/");
@@ -260,9 +248,6 @@ public class MeetingInProgressCommandTests
             MeetingInProgressUrl = "https://www.example.com/",
         };
         var connectorClientService = Substitute.For<IConnectorClientService>();
-        _ = connectorClientService
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>())
-            .Returns(new ResourceResponse());
         var dateTimeService = Substitute.For<IDateTimeService>();
         _ = dateTimeService
             .GetCurrentDateTime()
@@ -286,7 +271,7 @@ public class MeetingInProgressCommandTests
         // Assert
         _ = connectorClientService
             .DidNotReceive()
-            .SendActivityAsync("https://www.example.com/", Arg.Any<Activity>());
+            .SendActivityAsync(conversationReference, Arg.Any<Activity>());
         _ = qrCodeService
             .DidNotReceive()
             .CreateAsync("https://www.example.com/");